Stub for protoc-gen-jsontypes
The idea is to generate interfaces for the canonical JSON representation.

There is existing work in that direction, for example https://github.com/einride/protoc-gen-typescript-http and https://github.com/LLKennedy/protoc-gen-tsjson

There doesn't seem to be any plugin that supports the recommended JSON options (emit default values, etc.).

Not pursuing this idea ATM, but may pick it up again. This commit only adds a stub for a new plugin, nothing else.